Added rule type of fact=>fact and some descriptions on Fact creating view.

+++ ExpertSystem/ExpertSystem/Helpers/Solution.cs
@@ -60,43 +60,66 @@ namespace ExpertSystem
                     List<string> output = RPN.GetExpression(rule.Facts).Split().ToList();
                     output = RemoveEmptyEntries(output);
                     int i = 0;
-                    string a = output[i + 2];
+                    //string a = output[i + 2];
                     do
                     {
-                        if(output[i+2]!="^" && output[i+2]!="v")//Ищем знак действия
+                        if (output.Count > 2)
                         {
-                            i++;
-                            continue;
+                            if (output[i + 2] != "^" && output[i + 2] != "v")//Ищем знак действия
+                            {
+                                i++;
+                                continue;
+                            }
+                            else
+                            {
+                                if (output[i + 2] == "^")//Если конъюнкция
+                                {
+                                    if ((factsIdsList.Contains(output[i]) && factsIdsList.Contains(output[i + 1])) || (factsIdsList.Contains(output[i]) && output[i + 1] == "1True") || (factsIdsList.Contains(output[i + 1]) && output[i] == "1True"))//Если оба факта поданы в вопросе, то заменить 1True
+                                    {
+                                        output[i] = "1True";
+                                        output.RemoveRange(i + 1, 2);
+                                        i = 0;
+                                        continue;
+                                    }
+                                    else
+                                    {
+                                        break;//если кого то из фактов не было подано, то больше не рассматриваем правило
+                                    }
+                                }
+                                if (output[i + 2] == "v")//Если дизъюнкция
+                                {
+                                    if ((factsIdsList.Contains(output[i]) || factsIdsList.Contains(output[i + 1])) || (factsIdsList.Contains(output[i]) || output[i + 1] == "1True") || (factsIdsList.Contains(output[i + 1]) || output[i] == "1True"))//Если оба факта поданы в вопросе, то заменить 1True
+                                    {
+                                        output[i] = "1True";
+                                        output.RemoveRange(i + 1, 2);
+                                        i = 0;
+                                        continue;
+                                    }
+                                    else
+                                    {
+                                        break;//если кого то из фактов не было подано, то больше не рассматриваем правило
+                                    }
+                                }
+                            }
                         }
                         else
                         {
-                            if(output[i+2]=="^")//Если конъюнкция
+                            if(output.Count == 1)//Если только один факт
                             {
-                                if ((factsIdsList.Contains(output[i]) && factsIdsList.Contains(output[i + 1])) || (factsIdsList.Contains(output[i]) && output[i + 1] == "1True") || (factsIdsList.Contains(output[i+1]) && output[i] == "1True"))//Если оба факта поданы в вопросе, то заменить 1True
+                                if (factsIdsList.Contains(output[i]))
                                 {
                                     output[i] = "1True";
-                                    output.RemoveRange(i+1,2);
                                     i = 0;
                                     continue;
                                 }
                                 else
                                 {
-                                    break;//если кого то из фактов не было подано, то больше не рассматриваем правило
+                                    break;
                                 }
                             }
-                            if (output[i + 2] == "v")//Если дизъюнкция
+                            else
                             {
-                                if ((factsIdsList.Contains(output[i]) || factsIdsList.Contains(output[i + 1])) || (factsIdsList.Contains(output[i]) || output[i + 1] == "1True") || (factsIdsList.Contains(output[i + 1]) || output[i] == "1True"))//Если оба факта поданы в вопросе, то заменить 1True
-                                {
-                                    output[i] = "1True";
-                                    output.RemoveRange(i + 1, 2);
-                                    i = 0;
-                                    continue;
-                                }
-                                else
-                                {
-                                    break;//если кого то из фактов не было подано, то больше не рассматриваем правило
-                                }
+                                break;
                             }
                         }
                     }
